feat: Change steps in the main page
