+++ src/i18n/es.js
@@ -1089,14 +1089,14 @@ Con Doppler puedes generar Segmentos con intereses o características comunes, c
     },
     enter_button: 'Ingresar',
     go_to_login_button: 'Ingresar',
-    not_all_invoices_processed_legend: 'RReactivamos tu cuenta de Doppler para que puedas hacer el pago.',
+    not_all_invoices_processed_legend: 'Reactivamos tu cuenta de Doppler para que puedas hacer el pago.',
     not_all_invoices_processed_message: 'Hubo al menos un pago pendiente rechazado. Te contactaremos a {userEmail} para solucionarlo.',
     not_all_invoices_processed_title: 'Aún quedan pagos pendientes',
     payment_pending_message_line1: 'Confirmaremos tu pago en menos de 48 horas hábiles y te avisaremos a {userEmail}.',
     payment_pending_message_line2: 'Reactivamos tu cuenta para que puedas usar Doppler mientras esperamos que se confirme tu pago.',
     payment_pending_title: 'Tu pago con Mercado Pago está en proceso',
     rejected_payments_legend_1: `Prueba otra vez con una tarjeta internacional de <bold>Visa</bold>, <bold>American Express</bold> o <bold>Mastercard</bold>.`,
-    rejected_payments_legend_2: `Si necesitas ayuda, [contáctanos](${mailtoSupport})`,
+    rejected_payments_legend_2: `Si necesitas ayuda, [contáctanos](${mailtoSupport}).`,
     rejected_payments_message: ' Estos pagos pendientes fueron rechazados:',
     rejected_payments_title: 'Tu pago fue rechazado',
     title: 'Reproceso finalizado',
@@ -1104,7 +1104,7 @@ Con Doppler puedes generar Segmentos con intereses o características comunes, c
   updatePaymentMethod: {
     button_back: 'Volver',
     payment_method: {
-      save_continue_button: 'Continuar',
+      save_continue_button: 'Guardar y pagar',
       success_message: '¡Operación procesada con éxito!',
       title: 'Método de Pago',
       transfer: {
@@ -1126,7 +1126,7 @@ Con Doppler puedes generar Segmentos con intereses o características comunes, c
         error_column: 'Motivo que impidió el pago',
         invoice_column: 'Número de factura',
       },
-      payment_now_button: 'Pagar',
+      payment_now_button: 'Pagar todo',
       pending_amount_message: 'Total',
       success_message: '¡Operación procesada con éxito!',
       title: 'Pagos pendientes',
